Split directives for committeePageModule
