+++ app/components/appMainModule/committeePageModule/committeeContributorsPageController.js
@@ -1,0 +1,8 @@
+'use strict';
+
+function committeeContributorsPageController($scope, contributors) {
+  $scope.contributors = contributors;
+}
+
+module.exports = committeeContributorsPageController;
+
