+++ app/components/appMainModule/committeePageModule/committeeListing/committeeListingDirective.js
@@ -1,0 +1,19 @@
+/**
+ * committeeListingDirective.js
+ *
+ * Describe the committee details.
+ **/
+
+'use strict';
+
+function committeeListingDirective() {
+  return {
+    restrict: 'E',
+    scope: {
+      committee: '='
+    },
+    template: require('./committeeListing.html')
+  };
+}
+
+module.exports = committeeListingDirective;
