+++ app/components/appMainModule/committeePageModule/committeeListing/index.js
@@ -1,0 +1,18 @@
+/**
+ * committeeListing/index.js
+ *
+ * Describes a single political committee.
+ **/
+
+'use strict';
+
+require('../../../common/contributionAreaBreakdown');
+require('../../../common/contributionTypeBreakdown');
+
+var committeeListing = angular.module('committeeListing', [
+  'contributionAreaBreakdown',
+  'contributionTypeBreakdown'
+])
+  .directive('committeeListing', require('./committeeListingDirective'));
+
+module.exports = committeeListing;
