+++ app/components/appMainModule/committeePageModule/committeePageController.js
