--- app/components/appMainModule/committeePageModule/controllers/contributors.js
@@ -1,8 +1,0 @@
-'use strict';
-
-function contributorsController($scope, contributors) {
-  $scope.contributors = contributors;
-}
-
-module.exports = contributorsController;
-
